% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{age_symp_risk}
\alias{age_symp_risk}
\title{Function to calculate risk of an individual being symptomatic based on age.}
\usage{
age_symp_risk(age)
}
\arguments{
\item{age}{A number}
}
\value{
Estimated risk of being symptomatic
}
\description{
Function to calculate risk of an individual being symptomatic based on age.
}
