% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_mobility_functions.R
\name{county_file_download}
\alias{county_file_download}
\title{Downloading a crosswalk table of county and LAD codes}
\usage{
county_file_download(force_county = FALSE)
}
\arguments{
\item{force_county}{Logical. Should the data be downloaded even if there is
an existing download}
}
\description{
Downloading a crosswalk table of county and LAD codes
}
