% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{create_input}
\alias{create_input}
\title{Formatting data for infection model}
\usage{
create_input(micro_sim_pop, vars = NULL)
}
\arguments{
\item{micro_sim_pop}{The output of the spatial interaction model}

\item{vars}{Variables to be kept for use in the infection model}
}
\value{
A list of data to be used in the infection model
}
\description{
Formatting the output of the spatial interaction model for use in the
infection model and selecting which variables should be included.
}
\details{
BMI data recoded as:
"Not applicable" =  NA
"Underweight: less than 18.5" = 0
"Normal: 18.5 to less than 25" = 1
"Overweight: 25 to less than 30" = 2
 "Obese I: 30 to less than 35" = 3
"Obese II: 35 to less than 40" = 4
"Obese III: 40 or more" = 5
}
