% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{determine_removal}
\alias{determine_removal}
\title{Determines which individuals should be removed}
\usage{
determine_removal(df)
}
\arguments{
\item{df}{Input list of the function - output of the infection_length function}
}
\value{
idexes of individuals to be removed
}
\description{
Determines which individuals should be removed
}
