% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_mobility_functions.R
\name{format_gm}
\alias{format_gm}
\title{Format Google Mobility data}
\usage{
format_gm(gm_filt)
}
\arguments{
\item{gm_filt}{Output from the gm_filter function}
}
\value{
Long format data with the day and value for the
 residential change from baseline
}
\description{
Format the Google Mobility output to long format 
and looking at only the "residential_percent_change_from_baseline" data
}
