% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gam_cases}
\alias{gam_cases}
\title{Devon PHE cases}
\format{
A vector of integers, 103 in length:
}
\source{
\describe{PHE data}
}
\usage{
gam_cases
}
\description{
A data set containing smoothed PHE cases for Devon. 
Case numbers have been multiplied by 20 to get the desired 2% 
prevalence of COVID for this time-period
}
\keyword{datasets}
