% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_mobility_functions.R
\name{gm_file_download}
\alias{gm_file_download}
\title{Downloading Google Mobility Data}
\usage{
gm_file_download(force_gm = FALSE)
}
\arguments{
\item{force_gm}{Logical. Should the data be downloaded even if it 
already exists}
}
\description{
Downloading Google Mobility Data
}
