% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_mobility_functions.R
\name{gm_filter}
\alias{gm_filter}
\title{Filtering out the Google Mobility for the LAD or county}
\usage{
gm_filter(gm, lad_name, county_name)
}
\arguments{
\item{gm}{Google Mobility data - output of gm_file_download.}

\item{lad_name}{Name of LAD which we want Google Mobility data for.}

\item{county_name}{Name of county which we want Google Mobility data for, 
lad_name should be geographically within county_name.}
}
\value{
Google Mobility data for the given LAD or county
}
\description{
Filtering out the Google Mobility for the LAD or county
}
