% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_mobility_functions.R
\name{lad_file_download}
\alias{lad_file_download}
\title{Downloading crosswalk table of MSOA and LAD codes}
\usage{
lad_file_download(force_lad = FALSE)
}
\arguments{
\item{force_lad}{Logical. Should the data be downloaded even if there is
an existing download}
}
\description{
Downloading crosswalk table of MSOA and LAD codes
}
