% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_functions.R
\name{local_outbreak}
\alias{local_outbreak}
\title{increase probability of getting covid
for people in local outreak msoa}
\usage{
local_outbreak(df, msoa_infect, number_people = 100, risk_prob = 1)
}
\arguments{
\item{df}{data frame with individuals}

\item{msoa_infect}{msoa where the local outbreak occurs}

\item{number_people}{number of people at local outbreak event}

\item{risk_prob}{value to change probability of infection to for people involved in local outbreak}
}
\value{
updated dataframe of individuals with increased probability of getting infected for individuals at local outbreak
}
\description{
increase probability of getting covid
for people in local outreak msoa
}
