% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_mobility_functions.R
\name{lockdown_multiplier}
\alias{lockdown_multiplier}
\title{Converting the smoothed residential data into a multiplier for time outside}
\usage{
lockdown_multiplier(smth_res, pop)
}
\arguments{
\item{smth_res}{A numeric vector showing the (smoothed) proportion of time 
people have spent in residential locations}

\item{pop}{The output from the msoa_lad_code_matcher function}
}
\description{
Converting the smoothed residential data into a multiplier for time outside
}
