% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{logistic_map}
\alias{logistic_map}
\title{Function to map susceptibility}
\usage{
logistic_map(x)
}
\arguments{
\item{x}{factors that might affect susceptibility e.g smoking}
}
\value{
susceptibility
}
\description{
Function to map susceptibility
}
