% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{mortality_rate}
\alias{mortality_rate}
\title{Calculating mortality rate based on age and health risks}
\usage{
mortality_rate(df, obesity = 1.9)
}
\arguments{
\item{df}{The input list - the output from the create_input function}

\item{obesity}{The obesity mortality risk multiplier}
}
\value{
A list of data to be used in the infection model - 
with updated mortality risk
}
\description{
Calculating mortality rate based on age and health risks
}
