% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{mortality_risk}
\alias{mortality_risk}
\title{Calculating mortality rate based on age and health risks}
\usage{
mortality_risk(
  df,
  obesity = NULL,
  cvd = NULL,
  diabetes = NULL,
  bloodpressure = NULL
)
}
\arguments{
\item{df}{The input list - the output from the create_input function}

\item{obesity}{The obesity mortality risk multiplier}

\item{cvd}{The cardiovascular disease mortality risk multiplier}

\item{diabetes}{The disease mortality risk multiplier}

\item{bloodpressure}{The bloodpressure/hypertension mortality risk multiplier}
}
\value{
A list of data to be used in the infection model - 
with updated mortality risk
}
\description{
Calculating mortality rate based on age and health risks
}
