% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_mobility_functions.R
\name{msoa_lad_code_matcher}
\alias{msoa_lad_code_matcher}
\title{Matching the LAD level populations with their associated LAD and county names}
\usage{
msoa_lad_code_matcher(pop, lad_codes)
}
\arguments{
\item{pop}{An LAD level population with MSOA area codes}

\item{lad_codes}{Crosswalk of MSOA, LAD codes and county names}
}
\value{
The population with added LAD and county names
}
\description{
Matching the LAD level populations with their associated LAD and county names
}
