% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msoas}
\alias{msoas}
\title{High Risk MSOAS}
\format{
A data frame with 107 rows and 3 variables:
\describe{
  \item{area}{Name of MSOA}
  \item{risk}{Risk tertile, Low, Medium or High}
  ...
}
}
\source{
\describe{Calculated by Fiona Spooner # will put code on github}
}
\usage{
msoas
}
\description{
MSOAs in Devon and their risk level for COVID in terms of population density,
transport connectivity and wealth.
}
\keyword{datasets}
