% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{recalc_sympdays}
\alias{recalc_sympdays}
\title{Recalculates number of symptomatic and presymptomatic days remaining}
\usage{
recalc_sympdays(df)
}
\arguments{
\item{df}{Input list of the function - output of the removed function}
}
\value{
An updated version of the input list with the status updates for those
days left in stage = 0.
}
\description{
Recalculates number of symptomatic and presymptomatic days remaining
}
