% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{removed}
\alias{removed}
\title{Removes cases}
\usage{
removed(df, chance_recovery = 0.95)
}
\arguments{
\item{df}{Input list of the function - output of the infection_length function}

\item{chance_recovery}{Probability of an infected individual recovering}
}
\value{
An updated version of the input list with the status updates for those 
in their last day of being symptomatic/asymptomatic
}
\description{
For symptomatic individuals they have a 95% chance of recovery,
for asymptomatic they all recover.
}
