% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{removed_age}
\alias{removed_age}
\title{Removes cases based on age}
\usage{
removed_age(df)
}
\arguments{
\item{df}{Input list of the function - output of the infection_length function.
Must contain an age item.}
}
\value{
An updated version of the input list with the status updates for those 
in their last day of being symptomatic/asymptomatic
}
\description{
The survival rate of infected individuals decreases with age
}
