% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_mobility_functions.R
\name{residential_smoother}
\alias{residential_smoother}
\title{Smoothing the Google Mobility Residential data - there are some days with missing data}
\usage{
residential_smoother(residential_pcnt)
}
\arguments{
\item{residential_pcnt}{Google Mobility residential change 
from baseline data}
}
\value{
A smoothed set of values for the amount of time people spend 
in residential locations relative to the baseline
}
\description{
Smoothing the Google Mobility Residential data - there are some days with missing data
}
