% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{run_removal_recalc}
\alias{run_removal_recalc}
\title{Runs the pipeline for removing cases and stepping symptomatic and presymptomatic days}
\usage{
run_removal_recalc(df, chance_recovery = 0.95, seed = NULL)
}
\arguments{
\item{df}{Input list of the function - output of the infection_length function}

\item{chance_recovery}{Probability of an infected individual recovering}

\item{seed}{Seed for model run}
}
\value{
An updated version of the input list with the status updates for those
days left in stage = 0.
}
\description{
Runs the pipeline for removing cases and stepping symptomatic and presymptomatic days
}
