% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{run_status}
\alias{run_status}
\title{Run the infection model}
\usage{
run_status(
  pop,
  timestep = 1,
  current_risk_beta = 0.008,
  risk_cap = 5,
  seed_days = 10,
  exposed_dist = "weibull",
  exposed_mean = 2.56,
  exposed_sd = 0.72,
  presymp_dist = "weibull",
  presymp_mean = 2.3,
  presymp_sd = 0.35,
  infection_dist = "normal",
  infection_mean = 16,
  infection_sd = 3,
  asymp_rate = 0.7,
  chance_recovery = 0.95,
  output_switch = FALSE,
  rank_assign = FALSE,
  model_cases = NULL,
  w = NULL
)
}
\arguments{
\item{pop}{Microsim population - output of the 
spatial interaction model}

\item{timestep}{The day of the model run}

\item{current_risk_beta}{The multiplier of each individuals
hazard used to calculate probability of contracting COVID}

\item{risk_cap}{The value at which an individuals hazard is capped}

\item{seed_days}{The number of days which the model is seeded for -
based on PHE cases data}

\item{exposed_dist}{The distribution of the length of the exposed stage}

\item{exposed_mean}{The mean length of the exposed stage}

\item{exposed_sd}{The standard deviation of the length of 
the exposed stage}

\item{presymp_dist}{The distribution of the length of the 
presymptomatic stage}

\item{presymp_mean}{The mean length of the presymptomatic stage}

\item{presymp_sd}{The standard deviation of the length of the 
presymptomatic stage}

\item{infection_dist}{The distribution of the length of the symptomatic stage}

\item{infection_mean}{The mean length of the symptomatic stage}

\item{infection_sd}{The standard deviation of the length of the 
symptomatic stage}

\item{asymp_rate}{Percentage of infected people that are asymptomatic}

\item{chance_recovery}{Survival rate of those contracting COVID}

\item{output_switch}{Should the output be saved}

\item{rank_assign}{Should cases on all days be assigned by ranking 
current_risk}
}
\description{
Function that wraps all the other functions 
together to run the infection model each day
}
