% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{sympt_risk}
\alias{sympt_risk}
\title{Calculating risk of being symptomatic mortality rate based on age 
and health risks}
\usage{
sympt_risk(
  df,
  overweight_sympt_mplier = 1.46,
  cvd = NULL,
  diabetes = NULL,
  bloodpressure = NULL
)
}
\arguments{
\item{df}{The input list - the output from the create_input function}

\item{overweight_sympt_mplier}{The obesity risk multiplier for BMI > 40}

\item{cvd}{The cardiovascular disease mortality risk multiplier}

\item{diabetes}{The disease mortality risk multiplier}

\item{bloodpressure}{The bloodpressure/hypertension mortality risk multiplier}
}
\value{
A list of data to be used in the infection model - 
with updated mortality risk
}
\description{
Calculating risk of being symptomatic mortality rate based on age 
and health risks
}
