% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{vaccinate}
\alias{vaccinate}
\title{Vaccinate individuals on a given day according to the SMARTline flu jab survey}
\usage{
vaccinate(df)
}
\arguments{
\item{df}{Input list of the population - output of the recalc_sympdays function}
}
\value{
An updated population with approx 39% of the population vaccinated.
}
\description{
Vaccinate individuals on a given day according to the SMARTline flu jab survey
}
