% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{exp_cdf}
\alias{exp_cdf}
\title{Exponential distribution cdf, used to integrate intensity (0, inf) over time to a probability (0, 1)
 @param intensity current_risk from spatial interaction model
 @param dt timestep to integrate
 @return probability of infection
 @export}
\usage{
exp_cdf(intensity, dt)
}
\description{
Exponential distribution cdf, used to integrate intensity (0, inf) over time to a probability (0, 1)
 @param intensity current_risk from spatial interaction model
 @param dt timestep to integrate
 @return probability of infection
 @export
}
