% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{infection_length}
\alias{infection_length}
\title{Assigns the infection length of new cases}
\usage{
infection_length(
  df,
  exposed_dist = "weibull",
  exposed_mean = 3.2,
  exposed_sd = 1,
  presymp_dist = "weibull",
  presymp_mean = 3.2,
  presymp_sd = 1,
  infection_dist = "normal",
  infection_mean = 14,
  infection_sd = 2,
  asymp_rate = 0.5,
  overweight_sympt_mplier = 1
)
}
\arguments{
\item{df}{Input list of the function - output of an ____assign function}

\item{exposed_dist}{The distribution of the length of the exposed stage - options: weibull, lognormal, normal}

\item{exposed_mean}{The mean length of the exposed stage}

\item{exposed_sd}{The standard deviation of the length of the exposed stage}

\item{presymp_dist}{The distribution of the length of the presymptomatic stage - options: weibull, lognormal, normal}

\item{presymp_mean}{The mean length of the presymptomatic stage}

\item{presymp_sd}{The standard deviation of the length of the presymptomatic stage}

\item{infection_dist}{The distribution of the length of the symptomatic stage}

\item{infection_mean}{The mean length of the symptomatic stage}

\item{infection_sd}{The standard deviation of the length of the symptomatic stage}

\item{asymp_rate}{Percentage of infected people that are asymptomatic}

\item{overweight_sympt_mplier}{Multiplier to increase an overweight 
individuals chance of becoming symptomatic}
}
\value{
An updated version of the input list with the new cases having
infection lengths assigned
}
\description{
Each new case is assigned a number of days an individual is both
presymptomatic and symptomatic for.
}
